docs(examples): improve react-router example
