+++ examples/react/react-router/package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test:types": "tsc"
   },
   "dependencies": {
     "@tanstack/react-query": "^5.18.1",
@@ -20,7 +21,11 @@
     "sort-by": "^1.2.0"
   },
   "devDependencies": {
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "@types/sort-by": "^1.2.3",
     "@vitejs/plugin-react": "^4.2.1",
+    "typescript": "5.2.2",
     "vite": "^5.0.10"
   },
   "browserslist": {
