+++ examples/react/react-router/src/error-page.tsx
@@ -1,0 +1,23 @@
+import { useRouteError, isRouteErrorResponse } from 'react-router-dom'
+
+export default function ErrorPage() {
+  const error = useRouteError()
+  console.error(error)
+
+  let errorMessage = ''
+  if (isRouteErrorResponse(error)) {
+    errorMessage = error.statusText
+  } else if (error instanceof Error) {
+    errorMessage = error.message
+  }
+
+  return (
+    <div id="error-page">
+      <h1>Oops!</h1>
+      <p>Sorry, an unexpected error has occurred.</p>
+      <p>
+        <i>{errorMessage}</i>
+      </p>
+    </div>
+  )
+}
