+++ examples/react/react-router/src/index.tsx
@@ -7,7 +7,7 @@ import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import './index.css'
 
 import ErrorPage from './error-page'
-import Root, { loader as rootLoader, action as rootAction } from './routes/root'
+import Root, { loader as rootLoader } from './routes/root'
 import Contact, {
   loader as contactLoader,
   action as contactAction,
@@ -15,6 +15,7 @@ import Contact, {
 import EditContact, { action as editAction } from './routes/edit'
 import { action as destroyAction } from './routes/destroy'
 import Index from './routes/index'
+import NewContact, { action as newAction } from './routes/new'
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -30,40 +31,47 @@ const router = createBrowserRouter([
     element: <Root />,
     errorElement: <ErrorPage />,
     loader: rootLoader(queryClient),
-    action: rootAction(queryClient),
     children: [
       {
         index: true,
         element: <Index />,
       },
       {
+        path: 'contacts/new',
+        element: <NewContact />,
+        action: newAction(queryClient),
+        errorElement: <ErrorPage />,
+      },
+      {
         path: 'contacts/:contactId',
         element: <Contact />,
         loader: contactLoader(queryClient),
         action: contactAction(queryClient),
+        errorElement: <ErrorPage />,
       },
       {
         path: 'contacts/:contactId/edit',
         element: <EditContact />,
         loader: contactLoader(queryClient),
         action: editAction(queryClient),
+        errorElement: <ErrorPage />,
       },
       {
         path: 'contacts/:contactId/destroy',
         element: <EditContact />,
         action: destroyAction(queryClient),
-        errorElement: <div>Oops! There was an error.</div>,
+        errorElement: <ErrorPage />,
       },
     ],
   },
 ])
 
 const rootElement = document.getElementById('root')
-ReactDOM.createRoot(rootElement).render(
+ReactDOM.createRoot(rootElement!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <RouterProvider router={router} />
-      <ReactQueryDevtools position="bottom-right" />
+      <ReactQueryDevtools buttonPosition="bottom-right" />
     </QueryClientProvider>
   </React.StrictMode>,
 )
