+++ examples/react/react-router/src/routes/edit.tsx
@@ -1,20 +1,38 @@
-import * as React from 'react'
-import { Form, useLoaderData, redirect, useNavigate } from 'react-router-dom'
+import {
+  Form,
+  useLoaderData,
+  redirect,
+  useNavigate,
+  ActionFunctionArgs,
+} from 'react-router-dom'
 
-import { updateContact } from '../contacts'
+import { Contact, updateContact } from '../contacts'
+import { QueryClient, useSuspenseQuery } from '@tanstack/react-query'
+import { contactDetailQuery, loader } from './contact'
 
 export const action =
-  (queryClient) =>
-  async ({ request, params }) => {
+  (queryClient: QueryClient) =>
+  async ({ request, params }: ActionFunctionArgs) => {
     const formData = await request.formData()
     const updates = Object.fromEntries(formData)
+    if (!params.contactId) {
+      throw new Error('No contact ID provided')
+    }
     await updateContact(params.contactId, updates)
     queryClient.invalidateQueries({ queryKey: ['contacts'] })
     return redirect(`/contacts/${params.contactId}`)
   }
 
 export default function Edit() {
-  const contact = useLoaderData()
+  const { contactId } = useLoaderData() as Awaited<
+    ReturnType<ReturnType<typeof loader>>
+  >
+  const { data: contact } = useSuspenseQuery(contactDetailQuery(contactId))
+
+  return <ContactForm contact={contact} />
+}
+
+export function ContactForm({ contact }: { contact?: Contact }) {
   const navigate = useNavigate()
 
   return (
@@ -26,14 +44,14 @@ export default function Edit() {
           aria-label="First name"
           type="text"
           name="first"
-          defaultValue={contact.first}
+          defaultValue={contact?.first}
         />
         <input
           placeholder="Last"
           aria-label="Last name"
           type="text"
           name="last"
-          defaultValue={contact.last}
+          defaultValue={contact?.last}
         />
       </p>
       <label>
@@ -42,22 +60,21 @@ export default function Edit() {
           type="text"
           name="twitter"
           placeholder="@jack"
-          defaultValue={contact.twitter}
+          defaultValue={contact?.twitter}
         />
       </label>
       <label>
         <span>Avatar URL</span>
         <input
           placeholder="https://example.com/avatar.jpg"
-          aria-label="Avatar URL"
           type="text"
           name="avatar"
-          defaultValue={contact.avatar}
+          defaultValue={contact?.avatar}
         />
       </label>
       <label>
         <span>Notes</span>
-        <textarea name="notes" defaultValue={contact.notes} rows={6} />
+        <textarea name="notes" defaultValue={contact?.notes} rows={6} />
       </label>
       <p>
         <button type="submit">Save</button>
