+++ examples/react/react-router/src/routes/index.tsx
@@ -1,5 +1,3 @@
-import * as React from 'react'
-
 export default function Index() {
   return (
     <p id="zero-state">
@@ -9,9 +7,7 @@ export default function Index() {
       <a href="https://reactrouter.com/">
         the docs at reactrouter.com
       </a> and{' '}
-      <a href="https://tanstack.com/query/v4/docs/overview">
-        the docs at tanstack.com
-      </a>
+      <a href="https://tanstack.com/query/v5/docs/">the docs at tanstack.com</a>
       .
     </p>
   )
